move function to directory service
